move database helpers
