+++ README.md
@@ -38,6 +38,19 @@ DATABASE_URL = postgres://<user_name>:<password>@localhost:5432/<database_name>
 6. `npm run test` to run tests on helper functions, routes, and database queries.
 7. If you wish to edit the css, this project uses Sass (SCSS). Make sure you have [Sass installed](https://sass-lang.com/install), then go to the public folder and run `sass --watch scss:css`.
 
+---
+
+### File Structure
+
+📁 public - all publicly hosted files
+
+📁 src - all server side files
+    📁 database
+    📁 routes
+    📁 views
+
+📁 tests - all tests (run with tape)
+
 
 ---
