+++ public/css/style.css
@@ -560,7 +560,6 @@ a {
     content: "";
     position: absolute;
     display: none;
-    content: "";
     top: 0.5rem;
     left: 0.75rem;
     width: 0.5rem;
