+++ public/scss/_forms.scss
@@ -297,7 +297,6 @@
         content: "";
         position: absolute;
         display: none;
-        content: "";
         top: 0.5rem;
         left: 0.75rem;
         width: 0.5rem;
